fixed comparators

+++ src/PriorityComparator.java
@@ -6,7 +6,7 @@ public class PriorityComparator implements Comparator<Process>
 	@Override
 	public int compare(Process o1, Process o2) 
 	{
-		return o1.priority - o2.priority;
+		return (int) (o1.priority - o2.priority);
 	}
 
 }
