+++ src/SJFComparator.java
@@ -7,7 +7,7 @@ public class SJFComparator implements Comparator<Process>
 	@Override
 	public int compare(Process o1, Process o2)
 	{
-		return o1.burstValue - o2.burstValue;
+		return (int) (o1.burstValue - o2.burstValue);
 	}
 	
 }
